refactor(cli): carve daemon + cursor groups into cli/groups/
Move the two self-contained IDE/daemon groups out of the monolith into
cli/groups/daemon.py and cli/groups/cursor.py. __init__ imports them for
registration and re-exports the names. No behavior change (subcommands
intact; CLI tests pass).
